+++ common/src/main/java/mca/Config.java
@@ -42,6 +42,7 @@ public final class Config implements Serializable {
     //villager behavior
     public int chanceToHaveTwins = 2;
     public float marriageHeartsRequirement = 100;
+    public float engagementHeartsRequirement = 50;
     public int babyItemGrowUpTime = 24000;
     public int villagerMaxAgeTime = 192000;
     public int villagerMaxHealth = 20;