+++ common/src/main/java/mca/entity/ai/relationship/family/FamilyTreeNode.java
@@ -162,6 +162,7 @@ public final class FamilyTreeNode implements Serializable {
     public void updateMarriage(@Nullable Entity spouse, @Nullable MarriageState state) {
         this.spouse = spouse == null ? Util.NIL_UUID : spouse.getUuid();
         this.marriageState = state == null && spouse == null ? MarriageState.SINGLE : state;
+
         if (rootNode != null) {
             if (spouse != null) {
                 // ensure the family tree has an entry