+++ common/src/main/java/mca/item/EngagementRingItem.java
@@ -4,17 +4,17 @@ import mca.Config;
 import mca.entity.VillagerEntityMCA;
 import mca.entity.ai.Memories;
 import mca.entity.ai.Relationship;
+import mca.entity.ai.relationship.MarriageState;
 import mca.server.world.data.PlayerSaveData;
-import net.minecraft.item.Item;
 import net.minecraft.server.network.ServerPlayerEntity;
 
-public class WeddingRingItem extends TooltippedItem implements SpecialCaseGift {
-    public WeddingRingItem(Item.Settings properties) {
+public class EngagementRingItem extends TooltippedItem implements SpecialCaseGift {
+    public EngagementRingItem(Settings properties) {
         super(properties);
     }
 
     protected float getHeartsRequired() {
-        return Config.getInstance().marriageHeartsRequirement;
+        return Config.getInstance().engagementHeartsRequirement;
     }
 
     @Override
@@ -25,7 +25,7 @@ public class WeddingRingItem extends TooltippedItem implements SpecialCaseGift {
         boolean consume = false;
 
         if (villager.isBaby()) {
-            response = "interaction.marry.fail.isbaby";
+            response = "interaction.engage.fail.isbaby";
         } else if (Relationship.IS_PARENT.test(villager, player)) {
             response = "interaction.marry.fail.isparent";
         } else if (Relationship.IS_MARRIED.test(villager, player)) {
@@ -34,15 +34,16 @@ public class WeddingRingItem extends TooltippedItem implements SpecialCaseGift {
             response = "interaction.marry.fail.marriedtoother";
         } else if (villager.getRelationships().isEngaged() && !Relationship.IS_ENGAGED.test(villager, player)) {
             response = "interaction.marry.fail.engaged";
+        } else if (Relationship.IS_ENGAGED.test(villager, player)) {
+            response = "interaction.engage.fail.engaged";
         } else if (playerData.isMarried()) {
             response = "interaction.marry.fail.playermarried";
         } else if (memory.getHearts() < getHeartsRequired()) {
             response = "interaction.marry.fail.lowhearts";
         } else {
-            response = "interaction.marry.success";
-            playerData.marry(villager);
-            villager.getRelationships().marry(player);
-            villager.getVillagerBrain().modifyMoodValue(15);
+            response = "interaction.engage.success";
+            villager.getRelationships().getFamilyEntry().updateMarriage(player, MarriageState.ENGAGED);
+            villager.getVillagerBrain().modifyMoodValue(10);
             consume = true;
         }