+++ changelog.md
@@ -318,3 +318,5 @@
 # 7.3.3
 
 * Parents with same gender are now properly registered
+* Engagement rings now set the relationship to engaged
+  * Engaged villagers won't marry someone else